+++ src/containers/footer-lists/footer-lists.js
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux';
 
 import { Link } from 'react-router-dom';
 import Typography from '@material-ui/core/Typography';
-import { getCategoryURL } from '../../pages/home/categories/categories-list';
+import { getCategoryURL } from '../../pages/home/categories-list/categories-list';
 import { useStyles } from './footer-lists.styles';
 
 import {
