+++ src/containers/navbar-left/navbar-left.js
@@ -5,7 +5,7 @@ import Toolbar from '@material-ui/core/Toolbar';
 import { Link } from 'react-router-dom';
 import Typography from '@material-ui/core/Typography';
 import { useStyles } from './navbar-left.styles';
-import { getCategoryURL } from '../../pages/home/categories/categories-list';
+import { getCategoryURL } from '../../pages/home/categories-list/categories-list';
 
 import { LOGO } from '../../configs';
