+++ src/pages/home/categories-list/categories-list.style.js
