+++ src/pages/home/categories-list/category-item/category-item.style.js
