+++ src/pages/home/categories-list/category-item/index.js
