+++ src/redux/error/tests/error.reducer.test.js
@@ -22,7 +22,7 @@ describe('Error reducer test', () => {
         },
         {
           lang: 'en',
-          value: 'CategoriesList not found'
+          value: 'Categories not found'
         }
       ]
     };
